chore(suite): remove react-intersection-observer lib
