+++ packages/integration-tests/projects/suite-web/snapshots/firmware/firmware.test.ts/outdated notification banner.snap.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f9de53fb478622a6285c2cbefafbdb53ef536f3f17086adf601e961db7b40300
-size 38248
+oid sha256:85f88062e7938c2f87219400fb2a50363d143723501badeb7e120e4e8784b83b
+size 21785
