+++ packages/integration-tests/projects/suite-web/tests/settings/log.test.ts
@@ -15,8 +15,12 @@ describe('Log', () => {
     });
 
     it('there is a dropdown menu in settings that opens modal with application logs', () => {
-        cy.getTestElement('@settings/menu/dropdown').click();
-        cy.getTestElement('@settings/menu/log').click();
+        cy.getTestElement('@settings/menu/dropdown').click({
+            scrollBehavior: false,
+        });
+        cy.getTestElement('@settings/menu/log').click({
+            scrollBehavior: false,
+        });
         cy.getTestElement('@log/copy-button');
         cy.getTestElement('@log').matchImageSnapshot({ blackout: ['[data-test="@log/content"]'] });
