+++ packages/suite/package.json
@@ -42,7 +42,6 @@
         "react-dom": "17.0.1",
         "react-focus-lock": "^2.5.0",
         "react-hook-form": "6.15.7",
-        "react-intersection-observer": "^8.31.0",
         "react-intl": "^5.10.19",
         "react-markdown": "^7.1.1",
         "react-native-web": "^0.14.9",
